requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
future==1.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
minimalmodbus==2.1.1
pyserial==3.5
PyYAML==6.0.3
Werkzeug==3.1.3
